+++ CHANGELOG.md
@@ -7,12 +7,13 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### Enhancements
 
-* None.
+* Auxiliary files are excluded from backup by default.
 
 ### Bugfixes
 
 * Fix more cases where assigning an RLMArray property to itself would clear the
   RLMArray.
+* None.
 
 2.10.0 Release notes (2017-08-21)
 =============================================================